Add lifetimes to pointers.
I wanted a static guarantee that pointers could not be held across GC.
So:

1.  Add a lifetime bound on Ptr<'a>, representing a shared borrow of the
    `Storage`.
2.  Make `Storage::gc` receive `&mut self`.

These two rules ensure that no pointers are known to the user of
`Storage` when a GC runs - that would violate the lifetime bounds.

To handle roots, provide `return_roots` and `retrieve_roots` methods to
bracket the `gc` call. These "cache" the caller's roots, and retrieve
the translated / updated versions.

Also provide a `run_gc!` helper macro that handles the roots & GC
execution.