Debug, fixup trace stats
I was GCing properly (I think?) but wasn't tracing right: _again_
skipping Lisp objects based on the symbol index.

Maybe I should have a "walk" function instead of a "trace", so that I
don't make this mistake again.

...that said, I also need to implement interning myself, and implement
strings myself, to get to "all memory on the heap." So, maybe not right
away.