Switch store to fixed-size alternating arenas
This is a step towards `no_std`. The arenas are at the moment relatively small,
and dynamically allocated (`Box`ed) so they don't wind up on the stack
and don't have to deal with static-lifetime safety.

A fully `no_std` version would make them into static or start-time
allocated arenas; but we aren't doing that. Yet.